move AddCaseManagementAdapterClient to cms-adapter
