+++ partner-portal/src/API/Program.cs
@@ -5,6 +5,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Microsoft.OpenApi.Models;
+using Rsbc.Dmf.CaseManagement.Client;
 using Rsbc.Dmf.IcbcAdapter.Client;
 using Rsbc.Dmf.PartnerPortal.Api;
 using Rsbc.Dmf.PartnerPortal.Api.Model;
@@ -84,7 +85,7 @@ services.AddMemoryCache();
 
 // grpc clients
 services.AddDocumentStorageClient(builder.Configuration);
-services.AddCaseManagementAdapterClient(builder.Configuration);
+services.AddCaseManagementAdapterClient(builder.Configuration, loggerFactory);
 
 services.AddTransient<IUserService, UserService>();
 services.AddTransient<DocumentFactory>();
